add Rendering::IDrawable, fix bug,....

+++ Source/Engine/Source/ING/Utils/SmartArray/SmartArray.h
@@ -75,21 +75,21 @@ namespace ING {
 			 */
 		private:
 			T*												pData;
-			unsigned long									filledCount;
-			unsigned long									count;
+			size_t									filledCount;
+			size_t									count;
 
-			std::vector < std::unordered_map<unsigned long, unsigned long>* >	id2IndexMapVector;
-			std::vector < std::unordered_map<unsigned long, unsigned long>* >	index2IdMapVector;
+			std::vector < std::unordered_map<size_t, size_t>* >	id2IndexMapVector;
+			std::vector < std::unordered_map<size_t, size_t>* >	index2IdMapVector;
 
 		public:
 			T*												GetPData		() { return pData; }
 
-			unsigned long									GetCount		() { return count; }
+			size_t									GetCount		() { return count; }
 
-			unsigned long									GetFilledCount	() { return filledCount; }
+			size_t									GetFilledCount	() { return filledCount; }
 
-			std::vector < std::unordered_map<unsigned long, unsigned long>* > & GetId2IndexMapVector	() { return id2IndexMapVector; }
-			std::vector < std::unordered_map<unsigned long, unsigned long>* > & GetIndex2IdMapVector	() { return index2IdMapVector; }
+			std::vector < std::unordered_map<size_t, size_t>* > & GetId2IndexMapVector	() { return id2IndexMapVector; }
+			std::vector < std::unordered_map<size_t, size_t>* > & GetIndex2IdMapVector	() { return index2IdMapVector; }
 
 
 
@@ -97,18 +97,18 @@ namespace ING {
 			 *	Methods
 			 */
 		private:
-			void Allocate(unsigned long newFilledCount);
+			void Allocate(size_t newFilledCount);
 
 		public:
-			unsigned long Id2Index(unsigned long id);
-			unsigned long Index2Id(unsigned long index);
+			size_t Id2Index(size_t id);
+			size_t Index2Id(size_t index);
 
 		public:
 			void Clear();
-			void Add(T& data, unsigned long id);
-			void Erase(unsigned long id);
-			T&	 Get(unsigned long id);
-			T*	 GetDataPtr(unsigned long id);
+			void Add(T& data, size_t id);
+			void Erase(size_t id);
+			T&	 Get(size_t id);
+			T*	 GetDataPtr(size_t id);
 
 		};
 
@@ -129,9 +129,9 @@ namespace ING {
 		 *	Methods
 		 */
 		template<typename T>
-		unsigned long SmartArray<T>::Id2Index(unsigned long id) {
+		size_t SmartArray<T>::Id2Index(size_t id) {
 
-			for (std::unordered_map<unsigned long, unsigned long>*& id2IndexMap : id2IndexMapVector) {
+			for (std::unordered_map<size_t, size_t>*& id2IndexMap : id2IndexMapVector) {
 
 				if (id2IndexMap->find(id) != id2IndexMap->end()) {
 
@@ -146,9 +146,9 @@ namespace ING {
 		}
 
 		template<typename T>
-		unsigned long SmartArray<T>::Index2Id(unsigned long index) {
+		size_t SmartArray<T>::Index2Id(size_t index) {
 
-			for (std::unordered_map<unsigned long, unsigned long>*& index2IdMap : index2IdMapVector) {
+			for (std::unordered_map<size_t, size_t>*& index2IdMap : index2IdMapVector) {
 
 				if (index2IdMap->find(index) != index2IdMap->end()) {
 
@@ -163,9 +163,9 @@ namespace ING {
 		}
 
 		template<typename T>
-		void SmartArray<T>::Allocate(unsigned long newFilledCount) {
+		void SmartArray<T>::Allocate(size_t newFilledCount) {
 
-			unsigned long newCount = newFilledCount * 2;
+			size_t newCount = newFilledCount * 2;
 
 			if (newCount != 0) {
 
@@ -230,7 +230,7 @@ namespace ING {
 
 			filledCount = 0;
 
-			for (std::unordered_map<unsigned long, unsigned long>*& id2IndexMap : id2IndexMapVector) {
+			for (std::unordered_map<size_t, size_t>*& id2IndexMap : id2IndexMapVector) {
 
 				id2IndexMap->clear();
 
@@ -240,7 +240,7 @@ namespace ING {
 
 			id2IndexMapVector.clear();
 
-			for (std::unordered_map<unsigned long, unsigned long>*& index2IdMap : index2IdMapVector) {
+			for (std::unordered_map<size_t, size_t>*& index2IdMap : index2IdMapVector) {
 
 				index2IdMap->clear();
 
@@ -253,7 +253,7 @@ namespace ING {
 		}
 
 		template<typename T>
-		void SmartArray<T>::Add(T& data, unsigned long id) {
+		void SmartArray<T>::Add(T& data, size_t id) {
 
 			if (filledCount + 1 > count) {
 
@@ -266,13 +266,13 @@ namespace ING {
 
 			}
 
-			unsigned long index = filledCount - 1;
+			size_t index = filledCount - 1;
 
 			pData[index] = data;
 
 			bool isThereAnyId2IndexMapNotFull = false;
 
-			for (std::unordered_map<unsigned long, unsigned long>*& id2IndexMap : id2IndexMapVector) {
+			for (std::unordered_map<size_t, size_t>*& id2IndexMap : id2IndexMapVector) {
 
 				if (id2IndexMap->size() < id2IndexMap->max_size()) {
 
@@ -288,7 +288,7 @@ namespace ING {
 
 			if (!isThereAnyId2IndexMapNotFull) {
 
-				id2IndexMapVector.push_back(new std::unordered_map<unsigned long, unsigned long>());
+				id2IndexMapVector.push_back(new std::unordered_map<size_t, size_t>());
 
 				(*(id2IndexMapVector[id2IndexMapVector.size() - 1]))[id] = index;
 
@@ -298,7 +298,7 @@ namespace ING {
 
 			bool isThereAnyIndex2IdMapNotFull = false;
 
-			for (std::unordered_map<unsigned long, unsigned long>*& index2IdMap : index2IdMapVector) {
+			for (std::unordered_map<size_t, size_t>*& index2IdMap : index2IdMapVector) {
 
 				if (index2IdMap->size() < index2IdMap->max_size()) {
 
@@ -314,7 +314,7 @@ namespace ING {
 
 			if (!isThereAnyIndex2IdMapNotFull) {
 
-				index2IdMapVector.push_back(new std::unordered_map<unsigned long, unsigned long>());
+				index2IdMapVector.push_back(new std::unordered_map<size_t, size_t>());
 
 				(*(index2IdMapVector[index2IdMapVector.size() - 1]))[index] = id;
 
@@ -323,21 +323,21 @@ namespace ING {
 		}
 
 		template<typename T>
-		void SmartArray<T>::Erase(unsigned long id) {
+		void SmartArray<T>::Erase(size_t id) {
 
-			unsigned long index = Id2Index(id);
+			size_t index = Id2Index(id);
 
 			if (filledCount > 1) {
 
 				if (filledCount - 1 != index) {
 
-					unsigned long lastComponentId = Index2Id(filledCount - 1);
+					size_t lastComponentId = Index2Id(filledCount - 1);
 
 					pData[index] = pData[filledCount - 1];
 
 					//index2IdMap[index] = lastComponentId;
 
-					for (std::unordered_map<unsigned long, unsigned long>*& index2IdMap : index2IdMapVector) {
+					for (std::unordered_map<size_t, size_t>*& index2IdMap : index2IdMapVector) {
 
 						if (index2IdMap->find(index) != index2IdMap->end()) {
 
@@ -349,7 +349,7 @@ namespace ING {
 
 					}
 
-					for (std::unordered_map<unsigned long, unsigned long>*& id2IndexMap : id2IndexMapVector) {
+					for (std::unordered_map<size_t, size_t>*& id2IndexMap : id2IndexMapVector) {
 
 						if (id2IndexMap->find(lastComponentId) != id2IndexMap->end()) {
 
@@ -361,7 +361,7 @@ namespace ING {
 
 					}
 
-					for (std::unordered_map<unsigned long, unsigned long>*& index2IdMap : index2IdMapVector) {
+					for (std::unordered_map<size_t, size_t>*& index2IdMap : index2IdMapVector) {
 
 						if (index2IdMap->find(filledCount - 1) != index2IdMap->end()) {
 
@@ -373,7 +373,7 @@ namespace ING {
 
 					}
 
-					for (std::unordered_map<unsigned long, unsigned long>*& id2IndexMap : id2IndexMapVector) {
+					for (std::unordered_map<size_t, size_t>*& id2IndexMap : id2IndexMapVector) {
 
 						if (id2IndexMap->find(id) != id2IndexMap->end()) {
 
@@ -394,7 +394,7 @@ namespace ING {
 				}
 				else {
 
-					for (std::unordered_map<unsigned long, unsigned long>*& index2IdMap : index2IdMapVector) {
+					for (std::unordered_map<size_t, size_t>*& index2IdMap : index2IdMapVector) {
 
 						if (index2IdMap->find(index) != index2IdMap->end()) {
 
@@ -406,7 +406,7 @@ namespace ING {
 
 					}
 
-					for (std::unordered_map<unsigned long, unsigned long>*& id2IndexMap : id2IndexMapVector) {
+					for (std::unordered_map<size_t, size_t>*& id2IndexMap : id2IndexMapVector) {
 
 						if (id2IndexMap->find(id) != id2IndexMap->end()) {
 
@@ -444,7 +444,7 @@ namespace ING {
 			unsigned int index2IdMapVectorSize = index2IdMapVector.size();
 			for (unsigned int i = 0; i < index2IdMapVectorSize;) {
 
-				std::unordered_map<unsigned long, unsigned long>*& index2IdMap = index2IdMapVector[i];
+				std::unordered_map<size_t, size_t>*& index2IdMap = index2IdMapVector[i];
 
 				if (index2IdMap->size() == 0) {
 
@@ -464,7 +464,7 @@ namespace ING {
 			unsigned int id2IndexMapVectorSize = id2IndexMapVector.size();
 			for (unsigned int i = 0; i < id2IndexMapVectorSize;) {
 
-				std::unordered_map<unsigned long, unsigned long>*& id2IndexMap = id2IndexMapVector[i];
+				std::unordered_map<size_t, size_t>*& id2IndexMap = id2IndexMapVector[i];
 
 				if (id2IndexMap->size() == 0) {
 
@@ -482,14 +482,14 @@ namespace ING {
 		}
 
 		template<typename T>
-		T& SmartArray<T>::Get(unsigned long id) {
+		T& SmartArray<T>::Get(size_t id) {
 
 			return pData[Id2Index(id)];
 
 		}
 
 		template<typename T>
-		T* SmartArray<T>::GetDataPtr(unsigned long id) {
+		T* SmartArray<T>::GetDataPtr(size_t id) {
 
 			return pData + Id2Index(id);
